+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 ## [Unreleased]
 (Please put notes here)
 
+- Add Worklink service
 - Fix de/serialization of DynamoDB binary set attribute values
 - Change type for blob values from `Vec<u8>` to `Bytes`
+- Add DocDB service
 - Add Kafka service
 - Add Chime service
+- Add RDS Data service
 
 ## [0.38.0] - 2019-04-17